+++ jehibernate-testing/src/test/java/de/jexcellence/testfixture/MariaDbCrudIT.java
@@ -1,0 +1,9 @@
+package de.jexcellence.testfixture;
+
+import de.jexcellence.jehibernate.testing.JEHibernateTest;
+import de.jexcellence.jehibernate.testing.TestDatabase;
+
+/** CRUD scenario against MariaDB via Testcontainers (skipped without Docker). */
+@JEHibernateTest(database = TestDatabase.MARIADB, scanPackages = "de.jexcellence.testfixture")
+class MariaDbCrudIT extends AbstractCrudAcrossDatabasesIT {
+}